Refactor kubernetesCluster functions
